Add Boolean config include_internal
If { include_internal: true } is set in the yard-api
config, it will ignore @internal tags and create
documentation for those classes/methods anyway.

This will be useful for generating documentation sets
for internal use, as opposed to those intended for
public consumption which should omit any classes or
methods marked as "internal use only" with @internal.