Fix exception when generating JSON docs
If RouteHelper.api_methods_for_controller_and_action returns a route
with a verb that is a String instead of a Regexp, the get_route
method of the fulldoc/json method will throw an exception when
it tries to access route.verb.source.

This changes the route parsing of that method to use the RouteHelper
methods get_route_verb and get_route_path, which handle the types
better.