Fix broken CSS links on individual resources pages
In commit 4329ed1, a change was made that broke the generation
of the individual resources pages. The CSS links no longer point
to the correct location for the CSS files and the other embedded
links do not point to the correct locations.

This change backs out the change that is causing this problem.